On image uploads, use the index value the server responds with to assign the data-index variable and other identifies on DOM item to be added.
Noticed that when the user uploads image and it is using reader, the check is not picking up the duplicate filename because src attribute has the file content, not the actual filename.  Consider a new variable in the javascript (array or object) to store all the image filenames in the image container.